feat(gradedPicks): merge into exisitng coponent, fix bg color of non picked pick
