+++ packages/core/src/PickerButton/index.tsx
@@ -2,7 +2,8 @@ import React from "react";
 import { createUseStyles } from "react-jss";
 import classNames from "classnames";
 import color from "color";
-
+import iconCorrect from "./icon_correct.svg";
+import iconIncorrect from "./icon_incorrect.svg";
 import bolt from "./bolt.svg";
 
 import { DefaultTheme, PickerButtonProps } from "../types";
@@ -63,7 +64,7 @@ const useStyles = createUseStyles((theme: DefaultTheme) => ({
   bolt: {
     display: "none",
   },
-  isPick: {
+  hasIcon: {
     display: "block",
     "& img": {
       position: "relative",
@@ -91,6 +92,69 @@ const useStyles = createUseStyles((theme: DefaultTheme) => ({
       backgroundColor: theme.colors.purple.light,
     },
   },
+  nonPickedBar: {
+    display: "block",
+    position: "absolute",
+    top: 4,
+    left: 4,
+    width: "calc(100% - 8px)",
+    height: "calc(100% - 8px)",
+    borderRadius: 2,
+    zIndex: 1,
+    "& > div": {
+      display: "block",
+      position: "absolute",
+      top: 0,
+      left: 0,
+      width: 0,
+      height: "100%",
+      backgroundColor: theme.colors.grey.light,
+    },
+  },
+  isCorrect: {
+    border: `1px solid ${theme.colors.green.base}`,
+  },
+  isIncorrect: {
+    border: `1px solid ${theme.colors.red.base}`,
+  },
+  barCorrect: {
+    display: "block",
+    position: "absolute",
+    top: 4,
+    left: 4,
+    width: "calc(100% - 8px)",
+    height: "calc(100% - 8px)",
+    borderRadius: 2,
+    zIndex: 1,
+    "& > div": {
+      display: "block",
+      position: "absolute",
+      top: 0,
+      left: 0,
+      width: 0,
+      height: "100%",
+      backgroundColor: theme.colors.green.light,
+    },
+  },
+  barIncorrect: {
+    display: "block",
+    position: "absolute",
+    top: 4,
+    left: 4,
+    width: "calc(100% - 8px)",
+    height: "calc(100% - 8px)",
+    borderRadius: 2,
+    zIndex: 1,
+    "& > div": {
+      display: "block",
+      position: "absolute",
+      top: 0,
+      left: 0,
+      width: 0,
+      height: "100%",
+      backgroundColor: theme.colors.red.light,
+    },
+  },
   barActive: {
     "& > div": {
       animation: "$growRight 1s cubic-bezier(0.280, 0.520, 0.630, 0.990) 1",
@@ -107,8 +171,20 @@ const PickerButton: React.FC<PickerButtonProps> = ({
   showResult = false,
   result,
   isPick,
+  isCorrect,
+  isIncorrect,
 }) => {
   const classes = useStyles({ result });
+
+  const displayIcon = () => {
+    if (isCorrect) {
+      return iconCorrect;
+    } else if (isIncorrect) {
+      return iconIncorrect;
+    } else if (isPick) {
+      return bolt;
+    }
+  };
 
   return (
     <button
@@ -117,6 +193,8 @@ const PickerButton: React.FC<PickerButtonProps> = ({
         [classes.root]: true,
         [classes.result]: showResult,
         [classes.resultPicked]: isPick,
+        [classes.isCorrect]: isCorrect,
+        [classes.isIncorrect]: isIncorrect,
         [className]: className,
       })}
       onClick={onClick}
@@ -124,8 +202,11 @@ const PickerButton: React.FC<PickerButtonProps> = ({
     >
       <div
         className={classNames({
-          [classes.bar]: true,
+          [classes.bar]: isPick ? true : false,
+          [classes.nonPickedBar]: !isPick ? true : false,
           [classes.barActive]: showResult,
+          [classes.barCorrect]: isCorrect,
+          [classes.barIncorrect]: isIncorrect,
         })}
       >
         <div />
@@ -139,10 +220,10 @@ const PickerButton: React.FC<PickerButtonProps> = ({
             <div
               className={classNames({
                 [classes.bolt]: true,
-                [classes.isPick]: isPick,
+                [classes.hasIcon]: isPick || isCorrect || isIncorrect,
               })}
             >
-              <img src={bolt} aria-hidden role="presentation" />
+              <img src={displayIcon()} aria-hidden role="presentation" />
             </div>
             <div>{result}%</div>
           </div>
